take opg-flask-library, plug in simple incomes blueprint,  this is now working
