+++ docker-compose.yml
@@ -43,6 +43,7 @@ services:
     build:
       context: ./
       dockerfile: ./rest-api/docker/app/Dockerfile
+        #dockerfile: ./demo/Dockerfile
     depends_on:
       - postgres
       - feedbackdb
