+++ rest-api/docker/app/Dockerfile
@@ -1,16 +1,20 @@
-FROM 311462405659.dkr.ecr.eu-west-1.amazonaws.com/opg-flask-apps/rest-api:latest
+FROM python:3.9-alpine
 
 WORKDIR /restapi
 
+RUN apk update
+
+RUN apk add git openssh build-base
+
+RUN git clone --branch lpal-421-library-for-api https://github.com/ministryofjustice/opg-flask-apps.git
+
 COPY rest-api .
 
-RUN cat /restapi/app/extra_routes.py >> /restapi/app/routes.py
+RUN cd opg-flask-apps/flaskapi && pip install -e .
 
-RUN rm /restapi/app/extra_routes.py 
-
-RUN python setup.py build && python setup.py install
+RUN cd /restapi && python setup.py build && python setup.py install
 
 #ENV PYTHONPATH="/restapi/app:/restapi:$PYTHONPATH"
 
 EXPOSE 8005
-CMD ["gunicorn", "app:app", "--bind", "0.0.0.0:8005", "--log-file", "-"]
+CMD ["gunicorn", "opgfeedbackapi:api", "--bind", "0.0.0.0:8005", "--log-file", "-"]
